pyproject.toml
[tool.poetry]
name = "tdnsapilib"
version = "0.1.0-dev0"
description = "Api client library for technitium dns servers"
authors = ["Quien Sabe <qs5779@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
rest-api-client-framework = {git = "https://github.com/wtfo-guru/rest-api-client-framework.git", rev = "v0.1.1"}
pydantic = "^2.8.2"

[tool.poetry.group.test]
optional = false

[tool.poetry.group.test.dependencies]
mypy = "^1.9"
wemake-python-styleguide = "^0.19"
flake8-pytest-style = ">=1.6,<3.0"
doc8 = "^1.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest-cov = "^5.0.0"
testfixtures = "^8.3.0"
pytest-docker = "^3.1.1"

[tool.poetry.group.docs]
optional = false

[tool.poetry.group.docs.dependencies]
# sphinx = ">=5.2,<8.0"
# sphinx-autodoc-typehints = ">=1.24,<3.0"
m2r2 = "^0.3"
# tomli = "^2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
